03_Implementacao/server_projeto/requirements.txt
falcon
mysql-connector-python
falcon-cors
perlin-noise
requests